requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloud-sql-python-connector==1.17.0
cryptography==44.0.1
dnspython==2.7.0
frozenlist==1.5.0
google-auth==2.38.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
